requirements.txt
dropbox
requests
requests_html
lxml_html_clean
discord.py
sortedcontainers
cachetools
Pillow
texttable
humanize
pytz
python-dateutil
apscheduler
orjson
aiofile
polars
pyarrow
openpyxl
xlsx2csv
xlsxwriter
datetime_matcher
more_itertools
portion
sympy
bidict
feedparser
tweepy
exifread